components/Terminal/requirements.txt
csrbuilder==0.10.1
asn1crypto==1.2.0
components==0.0.6
pycrypto==2.6.1
pycryptodomex==3.9.4
pyOpenSSL==19.1.0
certvalidator==0.11.1
